servo: Merge #4979 - Audit ints and uints in script (from Ms2ger:audit-ints); r=Manishearth
Source-Repo: https://github.com/servo/servo
Source-Revision: 276f74b1ddec9dfa4cb053eb0802f95bd5ed6b66

UltraBlame original commit: 94a25931b4db8a27bd51a6e74574be0eb1e15514

+++ servo/components/script/dom/bindings/str.rs
@@ -34,7 +34,7 @@ impl ByteString {
     }
 
     
-    pub fn len(&self) -> uint {
+    pub fn len(&self) -> usize {
         let ByteString(ref vector) = *self;
         vector.len()
     }
